+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - [feat]: `.equals()` method;
+- [feat]: `.cross()` method;
+- [feat]: `.dot()` method;
 
 ## 0.5.0 (2022-11-27)